requirements.txt
langchain
langchain-community
langgraph
langchain-google-genai
python-dotenv
duckduckgo-search
unstructured
langchain-chroma
pytorch
faiss-cpu
